Fix bug preventing db rollback in Inputs, enable Camera Photo Action to capture stills with Camera Function
